+++ addons/scene_manager/scene_manager.gd
@@ -1,3 +1,4 @@
+@tool
 extends Node
 
 # consts
@@ -70,7 +71,8 @@ func _get_patterns() -> void:
 # set current scene and get patterns from `addons/scene_manager/shader_patterns` folder
 func _ready() -> void:
 	set_process(false)
-	_set_current_scene()
+	if not Engine.is_editor_hint():
+		_set_current_scene()
 	_get_patterns()
 
 # `speed` unit is in seconds